+++ packages/@livestore/node/src/worker-schema.ts
@@ -11,25 +11,16 @@ export const WorkerArgv = Schema.parseJson(
   }),
 )
 
-export const ExecutionBacklogItemExecute = Schema.TaggedStruct('execute', {
-  query: Schema.String,
-  bindValues: Schema.Any,
-})
-
 export const ExecutionBacklogItemMutate = Schema.TaggedStruct('mutate', {
   mutationEventEncoded: mutationEventSchemaEncodedAny,
   persisted: Schema.Boolean,
 })
 
 export const ExecutionBacklogItemTxn = Schema.TaggedStruct('txn', {
-  items: Schema.Union(ExecutionBacklogItemExecute, ExecutionBacklogItemMutate),
+  items: Schema.Union(ExecutionBacklogItemMutate),
 })
 
-export const ExecutionBacklogItem = Schema.Union(
-  ExecutionBacklogItemExecute,
-  ExecutionBacklogItemMutate,
-  ExecutionBacklogItemTxn,
-)
+export const ExecutionBacklogItem = Schema.Union(ExecutionBacklogItemMutate, ExecutionBacklogItemTxn)
 
 export type ExecutionBacklogItem = typeof ExecutionBacklogItem.Type